Reorganize core package exports.
